# Preallocate the histogram reader in matrix scanner